+++ NEWS.md
@@ -7,7 +7,7 @@
 ## Changes to functions
 
 * Make `tidy_stan()` for large `brmsfit`-objects (esp. with random effects) more efficient.
-* Better `print()`-method for `tidy_stan()`.
+* Better `print()`-method for `tidy_stan()`, `hdi()`, `rope()`, `icc()` and some other functions.
 * `link_inverse()` now also should return the link-inverse function for most (or some or all?) custom families of _brms_-models.
 
 ## New functions